prediction_server/docker/grpc_server/requirements.txt
grpcio==1.43.0
grpcio-tools==1.43.0
onnxruntime==1.10.0
uvicorn==0.16.0
boto3==1.20.41
numpy==1.22.1